Submit

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 /**
  * main - main block
- * Description: computes and prints even fibonacci number < 4,000,000
+ * Description: computes and prints even  number < 4,000,000
  * 5 below 1024 (excluded), followed by a new line
  * Return: 0
  */
 int main(void)
 {
 	int i = 0;
-	unsigned long int a = 0, b = 1, next = 0;
+	unsigned long long int a = 0, b = 1, next = 0;
 
 	while (next < 4000000)
 	{
